Refactor tests and update phone number handling
- Removed outdated test cases and added new tests for cleaning phone numbers from various formats, ensuring accurate normalization.
- Updated the `position_id` type in `SalesUser` struct from `Option<i32>` to `Option<u64>` for better compatibility.
- Cleaned up the main application structure by removing unnecessary test modules from `main.rs`.

+++ src/crud/users.rs
@@ -5,7 +5,7 @@ pub struct SalesUser {
     pub id: i32,
     pub telegram_id: Option<i64>,
     pub name: Option<String>,
-    pub position_id: Option<i32>,
+    pub position_id: Option<u64>,
     pub mtd_lead_count: i64,
 }
